Updated build config
