+++ src/style/index.css
@@ -1,20 +1,53 @@
-html, body {
-	height: 100%;
-	width: 100%;
-	padding: 0;
-	margin: 0;
-	background: #FAFAFA;
-	font-family: 'Helvetica Neue', arial, sans-serif;
-	font-weight: 400;
-	color: #444;
-	-webkit-font-smoothing: antialiased;
-	-moz-osx-font-smoothing: grayscale;
+body {
+    text-align: center;
+    background: #DFDFDF;
+    margin: 0 30px 100px 30px;
+    font: 14px/18px 'Lucida Grande', 'Segoe UI', sans-serif;
 }
 
-* {
-	box-sizing: border-box;
+h1 {
+    font: bold italic 50px Georgia, serif;
 }
 
-#app {
-	height: 100%;
+canvas {
+    display: block;
+    max-width: 800px;
+    background: white;
+    border-radius: 20px;
+    -moz-border-radius: 20px;
+    margin: 10px auto;
+}
+
+a {
+    color: black;
+}
+
+div {
+    margin: 30px auto;
+    text-align: left;
+    max-width: 800px;
+}
+
+.error {
+    display: block;
+    color: red;
+    font-size: 28px;
+    line-height: 30px;
+    padding: 30px;
+}
+
+p {
+    margin: 30px 0;
+    line-height: 20px;
+}
+
+.center {
+    text-align: center;
+}
+
+textarea {
+    display: none;
+    width: 75%;
+    height: 400px;
+    margin: 0 auto;
 }
