+++ src/template.html
@@ -36,13 +36,11 @@
 <head>
 	<meta charset="utf-8">
 	<title>
-		<% preact.title %>
+		fsm
 	</title>
 	<meta name="viewport" content="width=device-width,initial-scale=1">
 	<meta name="mobile-web-app-capable" content="yes">
 	<meta name="apple-mobile-web-app-capable" content="yes">
-	<link rel="apple-touch-icon" href="/assets/icons/apple-touch-icon.png">
-	<link rel="stylesheet" href="main.css">
 	<% preact.headEnd %>
 </head>
